dictionary now supports concurrent insertion, but is not acid, leaking on rollback, might produce version gap - essentially same semantics as sequences

+++ crates/engine/src/vm/volcano/scan/dictionary.rs
@@ -87,7 +87,7 @@ impl QueryNode for DictionaryScanNode {
 			}
 
 			if let Some(key) = DictionaryEntryIndexKey::decode(&entry.key) {
-				let entry_id = DictionaryEntryId::from_u128(key.id as u128, dict_def.id_type.clone())?;
+				let entry_id = DictionaryEntryId::from_u128(key.id, dict_def.id_type.clone())?;
 
 				let value: Value = from_bytes(&entry.row).map_err(|e| {
 					internal_error!("Failed to deserialize dictionary value: {}", e)
